mmc: sdhci: add quirk for max_discard calculation
The SDHCI driver by default specifies a parameter that causes the core
layer to calculate a max discard value which will be set on the mmc queue.
Unfortunately the value calculated because of this would be very small
compared to what comes in by default. As a result of this, any secure
discard kind of operations are very slow.

Add quirk so that the SDHCI host can use the default max_discard value.

CC: Krishna Konda <kkonda@codeaurora.org>
Signed-off-by: Georgi Djakov <gdjakov@mm-sol.com>