Input: synaptics-rmi4 - rename transport layer
The current RMI4 driver uses the term "physical" for two different logical
entities: the communications transport layer (I2C, SPI, and so on), and the
actual RMI4 device that is being communicated with.  Such usage makes the
code harder to understand due to confusion as to just which elements are
being referred to.

This patch renames the transport layer elements in order to eliminate this
confusion.  Much of the renaming was accomplished by the following Bash
script

  #!/bin/bash
  #
  # Update RMI4 driver transport layer naming.
  #

  files="rmi_bus.c rmi_bus.h rmi_driver.c rmi_driver.h rmi_f01.c rmi_i2c.c"
  cd drivers/input/rmi4

  for f in $files ; do
    echo $f
    sed -i.bak s/rmi_phys_device/rmi_transport_dev/g $f
    sed -i.bak s/rmi_phys_info/rmi_transport_info/g $f
    sed -i.bak "s/rmi_transport_dev \*phys/rmi_transport_dev \*xport/g" $f
    sed -i.bak "s/rmi_transport_dev \*rmi_phys/rmi_transport_dev \*xport/g" $f
    sed -i.bak "s/phys\([^i]\)/xport\1/g" $f
    sed -i.bak "s/->phys/->xport/g" $f
    sed -i.bak "s/register_physical_device/register_transport_device/g" $f
    sed -i.bak "s/physical_device_count/transport_device_count/g" $f
  done

although some changes proved easier to simply do by hand, particularly in the
comments.  Changes are confined strictly to the renaming, to keep the patch
relatively simple.

Signed-off-by: Christopher Heiny <cheiny@synaptics.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>