regulators: rpm: Implement RPM assisted disable
Some regulators are used to power e.g. PLLs that clocks the core we're
running on, so we can't turn them off directly; but we do want them off
when the core is powered down. To handle this the Qualcomm SoC provides
a means to specify a "sleep state" for RPM resources that can be
triggered by the hardware when the cores are brought down.

Resources of this type should be considered always-on while the core is
running and we can utilize the "sleep state" functionality to defer
disabling them until the core goes to sleep, if they are not used by
other peripherals at that time.

Signed-off-by: Bjorn Andersson <bjorn.andersson@sonymobile.com>