soc: qcom: smd: Expose functions needed to reset an edge
When a remote processor goes down the remoteproc driver needs a way to
tell the smd driver that the edge fell (as there is no explicit
indication of this in smd). So expose a function that can be called on
an edge pointer to close all open channels on that edge.

Signed-off-by: Bjorn Andersson <bjorn.andersson@sonymobile.com>