Input: synaptics-rmi4: Check for additional ACM registers appended to F11 data report
If a touchpad reports the F11 data40 register then this indicates that the
touchpad reports additional ACM (Accidental Contact Mitigation) data in
the F11 data registers. For HID touchpads these registers are appended to
the HID attention report. These additional bytes shift the position of the
F30 button data causing the driver to incorrectly report button state when
this functionality is present. This patch accounts for the additional data
in the report.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>