Input: synaptics-rmi4: Change the name of abs_size to attn_size
The abs_size variable is used to store the size of the F11 data in the HID
attention report when the HID transport is being used. Originally it was
called abs_size since attention reports only contain the abs registers.
However, some firmware is configured to report additional data in the
attention report so this patch gives this variable a more appropriate
name.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>