wcn36xx: Set SMD timeout to 10 seconds
After booting the wireless subsystem and uploading the NV blob to the
WCNSS_CTRL service the WiFi part continues to do things and will not
start servicing requests for another 2-5 seconds (measured).

The downstream code uses a timeout of 10 seconds for the communication
layer. By extending the wcn36xx timeout to match this we follows the
same flow for the boot procedure and can successfully configure WiFi
as wlan0 is registered.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>